+++ src/stores/shows/ShowsStore.ts
@@ -3,8 +3,8 @@ import ShowModel from './models/shows/ShowModel';
 import EpisodeModel from './models/episodes/EpisodeModel';
 import environment from 'environment';
 import groupBy from 'lodash.groupby';
-import IEpisodeTable from './computed/IEpisodeTable';
-import IEpisodeTableRow from './computed/IEpisodeTableRow';
+import IEpisodeTable from './models/computed/IEpisodeTable';
+import IEpisodeTableRow from './models/computed/IEpisodeTableRow';
 import dayjs from 'dayjs';
 import { initialResponseStatus, IResponseStatus } from '../../models/IResponseStatus';
 import { requestAction } from '../../utils/mobxUtil';
