+++ src/stores/shows/models/computed/IEpisodeTable.ts
