+++ src/stores/shows/models/computed/IEpisodeTableRow.ts
